[diffusion_labs] Move attention and res blocks under modules
[ghstack-poisoned]
