+++ torchmultimodal/diffusion_labs/models/adm_unet/adm.py
@@ -9,10 +9,8 @@ from typing import Callable, Dict, List, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import nn, Tensor
-from torchmultimodal.diffusion_labs.models.adm_unet.attention_block import (
-    adm_attn_block,
-)
-from torchmultimodal.diffusion_labs.models.adm_unet.res_block import (
+from torchmultimodal.diffusion_labs.modules.layers.attention_block import adm_attn_block
+from torchmultimodal.diffusion_labs.modules.layers.res_block import (
     adm_res_block,
     adm_res_downsample_block,
     adm_res_upsample_block,
