+++ torchmultimodal/diffusion_labs/modules/layers/res_block.py
